multi: hold timeout values in 'int' instead of 'long'
Since this code needs to work (identically) on systems with 32 bit longs
anyway, this might just as well always use 32 bit to save space and make
the code work the same everywhere.

+++ lib/multihandle.h
@@ -169,7 +169,7 @@ struct Curl_multi {
   /* timer callback and user data pointer for the *socket() API */
   curl_multi_timer_callback timer_cb;
   void *timer_userp;
-  long last_timeout_ms;        /* the last timeout value set via timer_cb */
+  int last_timeout_ms;        /* the last timeout value set via timer_cb */
   struct curltime last_expire_ts; /* timestamp of last expiry */
 
 #ifdef USE_WINSOCK
